# src/main/java/com/techwave/paymentservice/entity/CustomerAccountState.java
package com.techwave.paymentservice.entity;

public enum CustomerAccountState {
    DATA_REQUIRED,
    UNDER_REVIEW,
    ACCEPTED,
    REJECTED,
    NOT_REQUIRED,
    LAPSED,
    INACTIVE,
    CLOSED
}